Fixed internal segfault error

+++ src/adj_list.cpp
@@ -1338,7 +1338,7 @@ OutCursor *AdjList::get_outnbd_iter()
     uint64_t num_nodes = this->get_num_nodes();
     OutCursor *toReturn = new AdjOutCursor(get_out_adjlist_cursor(), session);
     toReturn->set_num_nodes(num_nodes);
-    toReturn->set_key_range({-1, num_nodes - 1});
+    toReturn->set_key_range({-1, (int64_t)num_nodes - 1});
     return toReturn;
 }
 
@@ -1347,7 +1347,7 @@ InCursor *AdjList::get_innbd_iter()
     uint64_t num_nodes = this->get_num_nodes();
     InCursor *toReturn = new AdjInCursor(get_in_adjlist_cursor(), session);
     toReturn->set_num_nodes(num_nodes);
-    toReturn->set_key_range({-1, num_nodes - 1});
+    toReturn->set_key_range({-1, (int64_t)num_nodes - 1});
     return toReturn;
 }
