+++ src/standard_graph.cpp
@@ -1133,7 +1133,7 @@ OutCursor *StandardGraph::get_outnbd_iter()
     uint64_t num_nodes = this->get_num_nodes();
     OutCursor *toReturn = new StdOutCursor(get_src_idx_cursor(), session);
     toReturn->set_num_nodes(num_nodes);
-    toReturn->set_key_range({-1, num_nodes - 1});
+    toReturn->set_key_range({-1, (int64_t)num_nodes - 1});
     return toReturn;
 }
 
@@ -1142,7 +1142,7 @@ InCursor *StandardGraph::get_innbd_iter()
     uint64_t num_nodes = this->get_num_nodes();
     InCursor *toReturn = new StdInCursor(get_dst_idx_cursor(), session);
     toReturn->set_num_nodes(num_nodes);
-    toReturn->set_key_range({-1, num_nodes - 1});
+    toReturn->set_key_range({-1, (int64_t)num_nodes - 1});
     return toReturn;
 }
